requirements.txt
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fonttools==4.62.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
joblib==1.5.3
kiwisolver==1.5.0
matplotlib==3.10.8
nltk==3.9.3
numpy==2.4.3
packaging==26.0
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-telegram-bot==22.6
regex==2026.2.28
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
threadpoolctl==3.6.0
tqdm==4.67.3
typing_extensions==4.15.0
wordcloud==1.9.6
